requirements.txt
selenium>=4.11.2
python-dateutil>=2.8.2
requests>=2.31.0
fastapi==0.95.2
uvicorn==0.22.0
